+++ app/layout.js
@@ -1,24 +1,10 @@
 import { GoogleTagManager } from "@next/third-parties/google";
 import { Inter } from "next/font/google";
-import dynamic from 'next/dynamic';
 import Footer from "./components/footer";
 import Navbar from "./components/navbar";
+import ClientComponents from "./client-components";
 import "./css/card.scss";
 import "./css/globals.scss";
-
-// Importaciones dinámicas para componentes que acceden al DOM
-const ToastContainer = dynamic(
-  () => import('react-toastify').then((mod) => mod.ToastContainer),
-  { ssr: false }
-);
-
-const ScrollToTop = dynamic(
-  () => import("./components/helper/scroll-to-top"),
-  { ssr: false }
-);
-
-// Importamos los estilos de react-toastify
-import "react-toastify/dist/ReactToastify.css";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -33,13 +19,11 @@ export default function RootLayout({ children }) {
         className={inter.className}
         suppressHydrationWarning={true}
       >
-        {/* Componentes renderizados solo en el cliente */}
-        <ToastContainer />
+        <ClientComponents />
         
         <main className="min-h-screen relative mx-auto px-6 sm:px-12 lg:max-w-[70rem] xl:max-w-[76rem] 2xl:max-w-[92rem] text-white">
           <Navbar />
           {children}
-          <ScrollToTop />
         </main>
         
         <Footer />
